+++ apps/web/src/main.tsx
@@ -6,6 +6,7 @@ import "@fontsource/ibm-plex-mono/latin-400.css";
 import "@fontsource/ibm-plex-mono/latin-500.css";
 import "./styles.css";
 import "./workspace-design.css";
+import "./workflow-polish.css";
 import { applyAppearance, readAppearance } from "./appearance";
 
 applyAppearance(readAppearance());